split rankings/models.py
separate files for Couple, Dancer
