+++ comps/heatlist/o2cm_heatlist.py
@@ -1,7 +1,8 @@
 import requests
 import html
 
-from rankings.models import Couple, Dancer
+from rankings.models.couple import Couple
+from rankings.models.dancer import Dancer
 from comps.models.heat import Heat
 from comps.models.heatlist_dancer import Heatlist_Dancer
 from comps.models.heatlist_error import Heatlist_Error
