+++ comps/views/couple_heats.py
@@ -2,7 +2,7 @@ from django.core.paginator import Paginator
 from django.shortcuts import render, get_object_or_404
 from comps.models.comp import Comp
 from comps.models.heat import Heat
-from rankings.models import Couple
+from rankings.models.couple import Couple
 from comps.filters import HeatFilter
