+++ rankings/management/commands/clear_models.py
@@ -1,5 +1,8 @@
 from django.core.management.base import BaseCommand
-from rankings.models import Dancer, Couple
+from rankings.models.couple import Couple
+from rankings.models.dancer import Dancer
+from rankings.models.elo_rating import EloRating
+
 
 class Command(BaseCommand):
     def handle(self, *args, **options):
