+++ rankings/models/__init__.py
@@ -1,0 +1,3 @@
+from .dancer import *
+from .couple import *
+from .elo_rating import *
